Initial check in Added 4 projects. +MVVM.Common +ProfilerCharts +UITestApp +WpfComponents -------------------------------------------------- MVVM.Common - Contains some classes that I've written several million times already. I finally decided to try to start compiling the most useful ones in this library. So far these include:     +IViewModel - Nothing in here yet, but this is meant to both extend the INotifyPropertyChanged interface and declare the most common ViewModel features like Name and, at some point probably some references for the validation system.     +ViewModelBase - This implements the IViewModel and INotifyPropertyChanged interfaces. I sometimes use a class called NotifyBase for this same thing, and eventually may extract that functionality from the ViewModelBase features, but for now this will handle both.
Eventually I plan to add some other of my mainstays like RequeryBase, CloseWindowCommand, and maybe a few other things, but for now, this will do.
-----------------------------------------------
ProfilerCharts - An application project that, initially conceived as a standalone profiling/charting application, and that's still what it is, but the plan is to eventually consolidate this and the UITestApp in some way that doesn't require the creation of multiple AppDomains and other crazy nonsense like it's currently doing.

Currently, this project can accept a comma delimited list of long int values as a startup parameter. These are loaded into the MainWindowViewModel and displayed as a simple LineSeries. That is not how it is executed by UITestApp however.
It's worth noting that although this is the only app that necessarily loads the WPFToolkit for its Charting libraries, because of some quirks within those libraries, the libraries don't load properly when this assembly is executed from anything other than it's own AppDomain. Which is why the UITestApp also references those libraries. Hopefully when I finish up some planned changes with the project startup order that will be fixed.
---------------------------------------------------
UITestApp - This application project loads any controls being tested into a window, generates a bunch of mock data to populate it with, and then loads and closes the window a bunch while logging the time it takes for each iteration. I made a concerted effort to isolate the Window.Show call chain so that it's only tracking how long it takes for WPF to make its way through the first rendering cycle including all of the measure and arrange stuff and any other layout engine stuff that is done the first time a control is shown on screen after creation.
The idea is to use it to benchmark both custom controls and various combinations of stock controls to try to pinpoint performance bottlenecks that can be avoided by making alterations to the defined layout.
--------------------------------------------------
WpfComponents - This is a controls library where I'll be defining all of UserControls and CustomControls I'll be making and testing with this solution.

TODO:
My next step is to reorder the project startup so that the ProfilerCharts App is launched first, gives the user a chance to select some tests to run, and set some parameters like number of tests, which controls will be tested, and how results should be compiled and displayed (allowing comparative data to be shown, or showing multiple successive runs of different versions of the same controls, etc.) and then it triggers the execution of the UITestApp test runs, and finally compiles and displays test results back in the ProfilerCharts app.