-Added the necessary DataTemplates to allow for multiple series to be displayed in the Chart simultaneously. -Refactored the test execution code so that it's less messy, a bit less fragile, and considerably less loathsome. The actual substance of that call chain has been better split between the TestSeries and TestItem portions. -I still need to make some decisions about exactly what ViewModels are going to be used to handle Test Results and Test Settings, as well as Series and Iterations of each. To be more clear, at the moment I'm considering: TestSeriesSettingsCollection TestSeriesResultsCollection Which, obviously would split responsibilities for the pre test setup and the post test results display. -The question that I still haven't fully sussed out is what the hierarchy looks like. Currently, the two areas of concern are represented by the following objects:
Settings
---------
-TestSeriesSettingsViewModel collection is contained in MainWindowViewModel.
-TestSettingsViewModel collection is contained in TestSeriesSettingsViewModel
-(No Bottom Level Settings Object)

Results
-------
-(No Top Level Results Object)
-ResultsViewModel collection contained in each TestSeriesSettingsViewModel
-ResultItemViewModel collection contained in each ResultsViewModel

The bottom level of the Results Hierarchy contains the data that represents a single datapoint, and therefore would probably be most aptly called a TestIterationViewModel, but since there aren't currently configurable parameters for how each individual iteration executes, there hasn't been a need for a corresponding TestIterationSettingsViewModel in the Settings Hierarchy. Which is really central to the question I mentioned earlier. There is something I find deeply objectionable about the fact that there isn't a perfectly parallel hierarchy. But at the moment, since the idea is to render a control a bunch of times so that a performance trend begins to reveal itself, I can't really make a convincing case for why a user would want to specify special conditions for any one iteration. But I'm not confident that there isn't a scenario where that would prove useful. So I don't want to build the settings hierarchy around a TestIterationSettingsViewModel because I don't want to have to deal with the extra complexity if I don't need to. But as I said, I think it really makes the taxonomy more understandable if both hierarchies have corresponding representations at each level. So, anyway, the point is, I'm eventually going to have to make a decision about this, but I haven't yet, so things are much harder to wrap your head around at the moment.

-Lastly, I've started to bump into some functionality gaps in the charting libraries that I may eventually have to address by creating my own fork of the WPFToolkit repo and just adding some things on my own, but I'll put that off as long as possible. In particular, I've had no success getting my normal datatemplating pattern to work in the charts. My normal method is to define multiple DataTemplates in the Resources of an ItemsControl, and having each specify a particular ViewModel type so that a collection can contain items of a variety of different types of data and each will have its own visual representation. This doesn't appear to work in this case as it typically does in standard controls. Instead there's just a single ItemTemplate property that may only be set once. At the moment, this is preventing me from having two different LineSeries types, one consisting of discrete datapoints and another that shows a horizontal line showing the average across all iterations. If I could use my preferred method, I could just define two different types of Results with a common interface and then throw them all in the same collection to be evaluated and templated according to their type. But, as I said, it doesn't seem to function in a way that makes that possible.